+++ src/routes/graphql/graphQLSchema.ts
@@ -2,7 +2,7 @@ import { GraphQLSchema } from "graphql";
 import queryTypes from "./queryTypes/queryTypes.js";
 import mutationTypes from "./mutationTypes/mutationTypes.js";
 
-const graphQLSchema = new GraphQLSchema({
+export const graphQLSchema = new GraphQLSchema({
   query: queryTypes,
   mutation: mutationTypes,
 });
