+++ src/routes/graphql/index.ts
@@ -3,7 +3,7 @@ import { createGqlResponseSchema, gqlResponseSchema } from './schemas.js';
 import { graphql, parse, validate } from 'graphql';
 import graphQLSchema from './graphQLSchema.js';
 import depthLimit from 'graphql-depth-limit';
-import managerLoader from './loaders/managerLoader.js';
+import manager from './loaders/manager.js';
 
 const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
   const { prisma } = fastify;
@@ -24,7 +24,7 @@ const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
       if (validationErrors.length > 0) {
         await reply.send({ errors: validationErrors });
       }
-      const loaders = managerLoader(prisma)
+      const loaders = manager(prisma)
 
       const res = await graphql({
         schema: graphQLSchema,
