+++ src/routes/graphql/loaders/profile.ts
@@ -1,0 +1,15 @@
+import { PrismaClient } from "@prisma/client";
+import DataLoader from "dataloader";
+
+const profile = (prisma: PrismaClient) => {
+  return new DataLoader(async (ids) => {
+    const profiles = await prisma.profile.findMany({
+      where: {
+        userId: { in: ids as string[] },
+      },
+    });
+    return ids.map((id) => profiles.find((profile) => profile.userId === id));
+  })
+};
+
+export default profile;
