--- src/routes/graphql/loaders/subscribedToUserLoader.ts
@@ -1,19 +1,0 @@
-import { PrismaClient } from "@prisma/client";
-import DataLoader from "dataloader";
-
-const subscribedToUserLoader = (prisma: PrismaClient) => {
-  return new DataLoader(async (ids) => {
-    const users = await prisma.user.findMany({
-      where: {
-        userSubscribedTo: {
-          some: {
-            authorId: { in: ids as string[] },
-          },
-        },
-      },
-    });
-    return ids.map(() => users);
-  })
-};
-
-export default subscribedToUserLoader;
