--- src/routes/graphql/mutationTypes/changePost/changePostObjectType.ts
@@ -1,16 +1,0 @@
-import { GraphQLInputObjectType } from "graphql";
-import { UUIDType } from "../../types/uuid.js";
-
-const changePostObjectType = new GraphQLInputObjectType({
-  name: 'ChangePostInput',
-  fields: () => ({
-    title: {
-      type: UUIDType,
-    },
-    content: {
-      type: UUIDType,
-    },
-  }),
-});
-
-export default changePostObjectType;
