--- src/routes/graphql/mutationTypes/changeUser/changeUserObjectType.ts
@@ -1,16 +1,0 @@
-import { GraphQLFloat, GraphQLInputObjectType } from "graphql";
-import { UUIDType } from "../../types/uuid.js";
-
-const changeUserObjectType = new GraphQLInputObjectType({
-  name: 'ChangeUserInput',
-  fields: () => ({
-    name: {
-      type: UUIDType,
-    },
-    balance: {
-      type: GraphQLFloat,
-    },
-  }),
-});
-
-export default changeUserObjectType;
