+++ src/routes/graphql/mutationTypes/create-post/create-post-object-type.ts
