+++ src/routes/graphql/mutationTypes/delete-post/delete-post.ts
@@ -1,0 +1,24 @@
+import { nonNullUUIDType } from "../../types/nonNullTypes.js";
+import { Post } from "@prisma/client";
+import Context from "../../types/context.js";
+
+const deletePost = {
+  deletePost: {
+    type: nonNullUUIDType,
+    args: {
+      id: {
+        type: nonNullUUIDType,
+      },
+    },
+    resolve: async (_source, args: Post, context: Context) => {
+      await context.prisma.post.delete({
+        where: {
+          id: args.id,
+        },
+      });
+      return args.id;
+    },
+  },
+};
+
+export default deletePost;
