+++ src/routes/graphql/mutationTypes/edit-user/edit-user-object-type.ts
@@ -1,0 +1,16 @@
+import { GraphQLFloat, GraphQLInputObjectType } from "graphql";
+import { UUIDType } from "../../types/uuid.js";
+
+const editUserObjectType = new GraphQLInputObjectType({
+  name: 'ChangeUserInput',
+  fields: () => ({
+    name: {
+      type: UUIDType,
+    },
+    balance: {
+      type: GraphQLFloat,
+    },
+  }),
+});
+
+export default editUserObjectType;
