+++ src/routes/graphql/queryTypes/userQuery/userObjectType.ts
@@ -1,6 +1,6 @@
 import { GraphQLObjectType, GraphQLFloat, GraphQLList } from "graphql";
 import { UUIDType } from "../../types/uuid.js";
-import IContext from "../../types/IContext.js";
+import Context from "../../types/context.js";
 import { User } from "@prisma/client";
 import {profileObjectType} from "../profileQuery/profileObjectType.js";
 import {postObjectTypeList} from "../postQuery/postObjectTypeList.js";
@@ -35,21 +35,21 @@ export const userObjectType = new GraphQLObjectType({
     profile: {
       type: profileObjectType as GraphQLObjectType,
       description: 'The profile',
-      resolve: async (source: User, _args, context: IContext) => {
+      resolve: async (source: User, _args, context: Context) => {
         return await context.loaders.profileLoader.load(source.id);
       },
     },
     posts: {
       type: postObjectTypeList,
       description: 'The posts',
-      resolve: async (source: User, _args: User, context: IContext) => {
+      resolve: async (source: User, _args: User, context: Context) => {
         return await context.loaders.postLoader.load(source.id);
       },
     },
     userSubscribedTo: {
       type: new GraphQLList(userObjectType),
       description: 'The userSubscribedTo',
-      resolve: async (source: IUserSubscribed, _args, context: IContext) => {
+      resolve: async (source: IUserSubscribed, _args, context: Context) => {
         if (source.userSubscribedTo) {
           const authorsId = source.userSubscribedTo.map((user) => user.authorId);
           return await context.loaders.userLoader.loadMany(authorsId);
@@ -60,7 +60,7 @@ export const userObjectType = new GraphQLObjectType({
     subscribedToUser: {
       type: new GraphQLList(userObjectType),
       description: 'The subscribedToUser',
-      resolve: async (source: IUserSubscribed, _args, context: IContext) => {
+      resolve: async (source: IUserSubscribed, _args, context: Context) => {
         if (source.subscribedToUser) {
           const subscribersId = source.subscribedToUser.map((user) => user.subscriberId);
           return await context.loaders.userLoader.loadMany(subscribersId);
